TrackOrDeckAttributes: Move TrackOrDeckAttributes and subclasses into separate files.
