+++ CMakeLists.txt
@@ -1371,6 +1371,10 @@ add_library(
   src/library/autodj/autodjprocessor.cpp
   src/library/autodj/dlgautodj.cpp
   src/library/autodj/dlgautodj.ui
+  src/library/autodj/track/deckattributes.cpp
+  src/library/autodj/track/fadeabletrackordeckattributes.cpp
+  src/library/autodj/track/trackattributes.cpp
+  src/library/autodj/track/trackordeckattributes.cpp
   src/library/banshee/bansheedbconnection.cpp
   src/library/banshee/bansheefeature.cpp
   src/library/banshee/bansheeplaylistmodel.cpp
