+++ src/library/autodj/autodjprocessor.cpp
@@ -22,164 +22,6 @@ constexpr double kMinimumTrackDurationSec = 0.2;
 constexpr bool sDebug = false;
 } // anonymous namespace
 
-TrackOrDeckAttributes::~TrackOrDeckAttributes() {
-}
-
-FadeableTrackOrDeckAttributes::FadeableTrackOrDeckAttributes()
-        : startPos(kKeepPosition),
-          fadeBeginPos(1.0),
-          fadeEndPos(1.0),
-          fadeDurationSeconds(0.0),
-          isFromDeck(false) {
-}
-
-FadeableTrackOrDeckAttributes::~FadeableTrackOrDeckAttributes() {
-}
-
-TrackAttributes::TrackAttributes(TrackPointer pTrack)
-        : m_pTrack(pTrack) {
-}
-
-TrackAttributes::~TrackAttributes() {
-}
-
-mixxx::audio::FramePos TrackAttributes::introStartPosition() const {
-    auto pIntro = m_pTrack->findCueByType(mixxx::CueType::Intro);
-    if (pIntro) {
-        return pIntro->getPosition();
-    }
-    return mixxx::audio::FramePos();
-}
-
-mixxx::audio::FramePos TrackAttributes::introEndPosition() const {
-    auto pIntro = m_pTrack->findCueByType(mixxx::CueType::Intro);
-    if (pIntro) {
-        return pIntro->getEndPosition();
-    }
-    return mixxx::audio::FramePos();
-}
-
-mixxx::audio::FramePos TrackAttributes::outroStartPosition() const {
-    auto pOutro = m_pTrack->findCueByType(mixxx::CueType::Outro);
-    if (pOutro) {
-        return pOutro->getPosition();
-    }
-    return mixxx::audio::FramePos();
-}
-
-mixxx::audio::FramePos TrackAttributes::outroEndPosition() const {
-    auto pOutro = m_pTrack->findCueByType(mixxx::CueType::Outro);
-    if (pOutro) {
-        return pOutro->getEndPosition();
-    }
-    return mixxx::audio::FramePos();
-}
-
-mixxx::audio::SampleRate TrackAttributes::sampleRate() const {
-    return m_pTrack->getSampleRate();
-}
-
-double TrackAttributes::playPosition() const {
-    return 0.0;
-}
-
-double TrackAttributes::rateRatio() const {
-    return 1.0;
-}
-
-mixxx::audio::FramePos TrackAttributes::trackEndPosition() const {
-    // Instead of actually loading the file, we simply infer
-    // the number of frames from duration and sample rate stored
-    // in the database. This isn't entirely accurate due to
-    // rounding errors, but more than accurate enough for
-    // estimating the remaining play time of the AutoDJ queue.
-    double approxNumSamples =
-            m_pTrack->getSampleRate() * m_pTrack->getDuration();
-
-    return mixxx::audio::FramePos(approxNumSamples);
-}
-
-DeckAttributes::DeckAttributes(int index,
-        BaseTrackPlayer* pPlayer)
-        : index(index),
-          group(pPlayer->getGroup()),
-          loading(false),
-          m_orientation(group, "orientation"),
-          m_playPos(group, "playposition"),
-          m_play(group, "play"),
-          m_repeat(group, "repeat"),
-          m_introStartPos(group, "intro_start_position"),
-          m_introEndPos(group, "intro_end_position"),
-          m_outroStartPos(group, "outro_start_position"),
-          m_outroEndPos(group, "outro_end_position"),
-          m_trackSamples(group, "track_samples"),
-          m_sampleRate(group, "track_samplerate"),
-          m_rateRatio(group, "rate_ratio"),
-          m_pPlayer(pPlayer) {
-    connect(m_pPlayer, &BaseTrackPlayer::newTrackLoaded,
-            this, &DeckAttributes::slotTrackLoaded);
-    connect(m_pPlayer, &BaseTrackPlayer::loadingTrack,
-            this, &DeckAttributes::slotLoadingTrack);
-    connect(m_pPlayer, &BaseTrackPlayer::playerEmpty,
-            this, &DeckAttributes::slotPlayerEmpty);
-    m_playPos.connectValueChanged(this, &DeckAttributes::slotPlayPosChanged);
-    m_play.connectValueChanged(this, &DeckAttributes::slotPlayChanged);
-    m_introStartPos.connectValueChanged(this, &DeckAttributes::slotIntroStartPositionChanged);
-    m_introEndPos.connectValueChanged(this, &DeckAttributes::slotIntroEndPositionChanged);
-    m_outroStartPos.connectValueChanged(this, &DeckAttributes::slotOutroStartPositionChanged);
-    m_outroEndPos.connectValueChanged(this, &DeckAttributes::slotOutroEndPositionChanged);
-    m_rateRatio.connectValueChanged(this, &DeckAttributes::slotRateChanged);
-}
-
-DeckAttributes::~DeckAttributes() {
-}
-
-void DeckAttributes::slotPlayChanged(double v) {
-    emit playChanged(this, v > 0.0);
-}
-
-void DeckAttributes::slotPlayPosChanged(double v) {
-    emit playPositionChanged(this, v);
-}
-
-void DeckAttributes::slotIntroStartPositionChanged(double v) {
-    emit introStartPositionChanged(this, v);
-}
-
-void DeckAttributes::slotIntroEndPositionChanged(double v) {
-    emit introEndPositionChanged(this, v);
-}
-
-void DeckAttributes::slotOutroStartPositionChanged(double v) {
-    emit outroStartPositionChanged(this, v);
-}
-
-void DeckAttributes::slotOutroEndPositionChanged(double v) {
-    emit outroEndPositionChanged(this, v);
-}
-
-void DeckAttributes::slotTrackLoaded(TrackPointer pTrack) {
-    emit trackLoaded(this, pTrack);
-}
-
-void DeckAttributes::slotLoadingTrack(TrackPointer pNewTrack, TrackPointer pOldTrack) {
-    //qDebug() << "DeckAttributes::slotLoadingTrack";
-    emit loadingTrack(this, pNewTrack, pOldTrack);
-}
-
-void DeckAttributes::slotPlayerEmpty() {
-    emit playerEmpty(this);
-}
-
-void DeckAttributes::slotRateChanged(double v) {
-    Q_UNUSED(v);
-    emit rateChanged(this);
-}
-
-TrackPointer DeckAttributes::getLoadedTrack() const {
-    return m_pPlayer != nullptr ? m_pPlayer->getLoadedTrack() : TrackPointer();
-}
-
 AutoDJProcessor::AutoDJProcessor(
         QObject* pParent,
         UserSettingsPointer pConfig,
