+++ src/library/autodj/track/deckattributes.h
@@ -1,0 +1,124 @@
+#pragma once
+
+#include "control/controlproxy.h"
+#include "engine/channels/enginechannel.h"
+#include "library/autodj/track/fadeabletrackordeckattributes.h"
+
+class BaseTrackPlayer;
+
+/// Exposes the attributes of the track loaded in a certain player deck
+class DeckAttributes : public FadeableTrackOrDeckAttributes {
+    Q_OBJECT
+  public:
+    DeckAttributes(int index, BaseTrackPlayer* pPlayer);
+    virtual ~DeckAttributes();
+
+    bool isLeft() const {
+        return m_orientation.get() == static_cast<double>(EngineChannel::LEFT);
+    }
+
+    bool isRight() const {
+        return m_orientation.get() == static_cast<double>(EngineChannel::RIGHT);
+    }
+
+    bool isPlaying() const {
+        return m_play.toBool();
+    }
+
+    void stop() {
+        m_play.set(0.0);
+    }
+
+    void play() {
+        m_play.set(1.0);
+    }
+
+    double playPosition() const override {
+        return m_playPos.get();
+    }
+
+    void setPlayPosition(double playpos) {
+        m_playPos.set(playpos);
+    }
+
+    bool isRepeat() const {
+        return m_repeat.toBool();
+    }
+
+    void setRepeat(bool enabled) {
+        m_repeat.set(enabled ? 1.0 : 0.0);
+    }
+
+    mixxx::audio::FramePos introStartPosition() const override {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_introStartPos.get());
+    }
+
+    mixxx::audio::FramePos introEndPosition() const override {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_introEndPos.get());
+    }
+
+    mixxx::audio::FramePos outroStartPosition() const override {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_outroStartPos.get());
+    }
+
+    mixxx::audio::FramePos outroEndPosition() const override {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_outroEndPos.get());
+    }
+
+    mixxx::audio::SampleRate sampleRate() const override {
+        return mixxx::audio::SampleRate::fromDouble(m_sampleRate.get());
+    }
+
+    mixxx::audio::FramePos trackEndPosition() const override {
+        return mixxx::audio::FramePos::fromEngineSamplePosMaybeInvalid(m_trackSamples.get());
+    }
+
+    double rateRatio() const override {
+        return m_rateRatio.get();
+    }
+
+    TrackPointer getLoadedTrack() const override;
+
+  signals:
+    void playChanged(DeckAttributes* pDeck, bool playing);
+    void playPositionChanged(DeckAttributes* pDeck, double playPosition);
+    void introStartPositionChanged(DeckAttributes* pDeck, double introStartPosition);
+    void introEndPositionChanged(DeckAttributes* pDeck, double introEndPosition);
+    void outroStartPositionChanged(DeckAttributes* pDeck, double outtroStartPosition);
+    void outroEndPositionChanged(DeckAttributes* pDeck, double outroEndPosition);
+    void trackLoaded(DeckAttributes* pDeck, TrackPointer pTrack);
+    void loadingTrack(DeckAttributes* pDeck, TrackPointer pNewTrack, TrackPointer pOldTrack);
+    void playerEmpty(DeckAttributes* pDeck);
+    void rateChanged(DeckAttributes* pDeck);
+
+  private slots:
+    void slotPlayPosChanged(double v);
+    void slotPlayChanged(double v);
+    void slotIntroStartPositionChanged(double v);
+    void slotIntroEndPositionChanged(double v);
+    void slotOutroStartPositionChanged(double v);
+    void slotOutroEndPositionChanged(double v);
+    void slotTrackLoaded(TrackPointer pTrack);
+    void slotLoadingTrack(TrackPointer pNewTrack, TrackPointer pOldTrack);
+    void slotPlayerEmpty();
+    void slotRateChanged(double v);
+
+  public:
+    int index;
+    QString group;
+    bool loading; // The data is inconsistent during loading a deck
+
+  private:
+    ControlProxy m_orientation;
+    ControlProxy m_playPos;
+    ControlProxy m_play;
+    ControlProxy m_repeat;
+    ControlProxy m_introStartPos;
+    ControlProxy m_introEndPos;
+    ControlProxy m_outroStartPos;
+    ControlProxy m_outroEndPos;
+    ControlProxy m_trackSamples;
+    ControlProxy m_sampleRate;
+    ControlProxy m_rateRatio;
+    BaseTrackPlayer* m_pPlayer;
+};
