+++ src/library/autodj/track/fadeabletrackordeckattributes.cpp
@@ -1,0 +1,12 @@
+#include "library/autodj/track/fadeabletrackordeckattributes.h"
+
+#include "library/autodj/autodjconstants.h"
+#include "moc_fadeabletrackordeckattributes.cpp"
+
+FadeableTrackOrDeckAttributes::FadeableTrackOrDeckAttributes()
+        : startPos(AutoDJConstants::kKeepPosition),
+          fadeBeginPos(1.0),
+          fadeEndPos(1.0),
+          fadeDurationSeconds(0.0),
+          isFromDeck(false) {
+}
