+++ src/library/autodj/track/trackordeckattributes.cpp
@@ -1,0 +1,6 @@
+#include "library/autodj/track/trackordeckattributes.h"
+
+#include "moc_trackordeckattributes.cpp"
+
+TrackOrDeckAttributes::~TrackOrDeckAttributes() {
+}
